requirements[1].txt
flask
pdfplumber
langchain
langchain-community
langchain-google-genai
chromadb
python-dotenv
google-generativeai
